move github link into header
