+++ src/components/Header/Header.scss
@@ -164,3 +164,15 @@
 .header--scroll {
   position: relative;
 }
+
+.github-corner {
+  display: none;
+  position: absolute;
+  top: 0;
+  right: 0;
+  z-index: 100;
+
+  @include desktop {
+    display: block;
+  }
+}
