+++ src/components/Layout/Layout.scss
@@ -16,16 +16,3 @@
     }
   }
 }
-
-.github-corner {
-  display: none;
-  position: fixed;
-  top: 0;
-  right: 0;
-  mix-blend-mode: multiply;
-  z-index: 100;
-
-  @include desktop {
-    display: block;
-  }
-}
