+++ src/components/Layout/index.js
@@ -5,8 +5,6 @@ import Footer from '../Footer';
 import Notification from '../Notification';
 import './Layout.scss';
 
-import githubCorner from './github-corner.svg';
-
 let Layout = ({ children, location }) => {
   // in some cases it's useful to mark the root object with a class to identify the current page, in case
   // some pages want to make modifications to the layout object
@@ -21,24 +19,9 @@ let Layout = ({ children, location }) => {
       <Header />
       {children}
       <Footer />
-
-      <GithubCorner />
     </div>
   );
 };
 Layout = withRouter(Layout);
 
 export default Layout;
-
-function GithubCorner() {
-  return (
-    <a
-      href="https://github.com/AlexsLemonade"
-      target="_blank"
-      rel="nofollow noopener noreferrer"
-      className="github-corner"
-    >
-      <img src={githubCorner} />
-    </a>
-  );
-}
